Create lead package and move common classes and functions over
